Fix a bug in cvmix_WSwSBF init mode
When initializing the temperature and salinity profiles from their
surface values and constant gradients, the values at the first level
were assumed to be the surface values, which is not consistent with
how Brunt-Vaisala frequency is computed in ocn_diagnostics().
Now using the values at the midpoint of the first vertical grid.